requirements.txt
# This file is used by pip to install required python packages
# Usage: pip install -r requirements.txt

# REST/HTTP
Flask==2.0.1
requests==2.24.0
fastapi==0.65.3
uvicorn==0.14.0

PyYAML==5.4.1

# Persistence
#tinydb==4.4.0
#mysql-connector-python==8.0.22

# Retry
#tenacity==6.2.0